# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Back End Coursework API
  description: A simple e-commerce API.
  contact:
    name: Joseph Lees
    email: leesj4@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Development server
...
paths:
...
  /users/{id}/addresses:
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the user
    get:
      tags:
        - Addresses
      description:
        Get all addresses for a user.
      responses:
        '200':
          description: View all addresses for a user.
        '403':
          description: Non-admin users can only view their own addresses.
        '404':
          description: No addresses found for the user.
        '500':
          description: Internal server error.
    post:
      tags:
        - Addresses
      description:
        Create a new address for the user.
      requestBody:
        description: new address information
        required: true
        content:
          application/json:
            schema:
              $ref: ./address.json#/definitions/address
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                $ref: ./address.json#/definitions/addressCreated
        '403':
          description: Non-admin users can only create addresses for themselves, or the user already has a registered address.
        '429':
          description: User has made a post request too recently.
        '500':
          description: Internal server error.
  /users/{id}/addresses/{addressId}:
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the user
        - in: path
          name: addressId
          schema:
            type: integer
          required: true
          description: Numeric id of the address
    get:
      tags:
        - Addresses
      description:
        Get address by id.
      responses:
        '200':
          description: Address record retrieved.
        '403':
          description: Non-admin users can only view their own addresses.
        '404':
          description: Address was not found in the DB.
        '500':
          description: Internal server error.
    put:
      tags:
        - Addresses
      description:
        Update address by id.
      requestBody:
        description: updated address information
        required: true
        content:
          application/json:
            schema:
              $ref: ./address.json#/definitions/addressUpdate
      responses:
        '200':
          description: Address was updated successfully.
          content:
            application/json:
              schema:
                $ref: ./address.json#/definitions/addressUpdated
        '403':
          description: Non-admin users can only update their own addresses.
        '404':
          description: Address was not found in the DB.
        '500':
          description: Internal server error.
    delete:
      tags:
        - Addresses
      description:
        Delete address by id.
      responses:
        '204':
          description: Address was deleted successfully.
        '403':
          description: Non-admin users can only delete their own addresses.
        '404':
          description: Address was not found in the DB.
        '500':
          description: Internal server error.
...
  /categories:
    summary: Represents an individual category
    description: >
      This resource represents an individual category in the system.
      Each category is identified by a numeric `id`.
    post:
      tags:
        - Categories
      description:
        Create a new category.
      requestBody:
        description: new category information
        required: true
        content:
          application/json:
            schema:
              $ref: ./category.json#/definitions/category
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: ./category.json#/definitions/categoryCreated
        '403':
          description: Non-admin users can't create categories.
        '500':
          description: Internal server error.
    get:
      tags:
        - Categories
      description:
        Get all categories.
      responses:
        '200':
          description: View of all category records
        '403':
          description: Non-admin users can't view categories.
        '404':
          description: No categories found in the DB.
        '500':
          description: Internal server error.
  /categories/{id}:
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the category
    get:
      tags:
        - Categories
      description:
        Get category by id.
      responses:
        '200':
          description: View of a category record.
        '403':
          description: Non-admins cannot view categories.
        '404':
          description: Category was not found in the DB.
        '500':
          description: Internal server error.
    put:
      tags:
        - Categories
      description:
        Update category by id.
      requestBody:
        description: updated category information
        required: true
        content:
          application/json:
            schema:
              $ref: ./category.json#/definitions/categoryUpdate
      responses:
        '200':
          description: Category was updated successfully.
          content:
            application/json:
              schema:
                $ref: ./category.json#/definitions/categoryUpdated
        '403':
          description: Non-admin users can't update categories.
        '404':
          description: Category was not found in the DB.
        '500':
          description: Internal server error.
    delete:
      tags:
        - Categories
      description:
        Delete category by id.
      responses:
        '204':
          description: Category was deleted successfully.
        '403':
          description: Non-admin users can't delete categories.
        '404':
          description: Category was not found in the DB.
        '500':
          description: Internal server error.
...